support opening database read/write
This commit introduces a type class (not exported) for database
mode, and databaseOpen changes to support opening the database RO or
RW depending on the type parameter.  In case of computations
involving RO-only or RW-only operations, the type will be deduced.
Otherwise a type annotation is needed to tell databaseOpen whether
a RO or RW database handle is wanted.