add DatabaseMode type param to Query, Thread, Message
We need to ensure that message routines that are only valid for
messages derived from a read/write database cannot be applied to
messages that were not derived from a read/write database.

Add a DatabaseMode type parameter to the Query, Thread and Message
data types, using the type system to tag messages with the mode of
the database they came from.