avoid use of ForeignPtr for notmuch_tags_t
Tags are always eagerly copied from the libnotmuch-managed memory,
within a 'withForeignPtr' of the owner of the notmuch_tags_t object.
So don't bother wrapping notmuch_tags_t objects as a ForeignPtr and
don't 'deatchPtr', reducing our overhead and allowing the memory to
be freed when the owner object (e.g. message, thread) is destroyed.

This saves a bit of time and a significant number of allocations.
For an invocation of hs-notmuch-tags-count with 161474 messages and
183379 tags the improvement is:

  total time  =        8.34 secs   (8335 ticks @ 1000 us, 1 processor)
  total alloc = 244,189,264 bytes  (excludes profiling overheads)

versus:

  total time  =        8.92 secs   (8918 ticks @ 1000 us, 1 processor)
  total alloc = 315,237,824 bytes  (excludes profiling overheads)