lazily iterate threads and messages
Iterating a notmuch_(messages|threads|tags)_t happens all at once,
even if the result is not used.  For example, a modified
hs-notmuch-tag-count that only reads the tags from 'take 10 msgs'
still constructs the whole message list.  Searching '*' on a
database with 161474 messages results in:

  total time  =        1.79 secs   (1795 ticks @ 1000 us, 1 processor)
  total alloc =  59,500,568 bytes  (excludes profiling overheads)

                                    individual      inherited
  COST CENTRE            entries  %time %alloc   %time %alloc
   message_get_tags          10    0.0    0.0     0.0    0.0
   ...
     messagesToList           0    1.3   89.0    96.3   99.9
      detachPtr          161474   17.5    0.0    17.5    0.0

Add 'lazyPtrToList' which uses unsafeInterleaveIO to suspend reading
the iterator, and use this function for messages and threads.  Tag
lists are still read eagerly.  With this change the profiler shows:

  total time  =        0.07 secs   (68 ticks @ 1000 us, 1 processor)
  total alloc =      79,960 bytes  (excludes profiling overheads)

                                    individual      inherited
  COST CENTRE            entries  %time %alloc   %time %alloc
     message_get_tags        10    0.0    0.5     0.0    9.5
  ...
     messagesToList           0    0.0    0.0     0.0    6.2
         detachPtr           10    0.0    0.0     0.0    0.0

When iterating the full 161474 messages, the performance changes
from:

     357,798,040 bytes allocated in the heap
     334,331,640 bytes copied during GC
      68,431,240 bytes maximum residency (9 sample(s))
         385,248 bytes maximum slop
             148 MB total memory in use (0 MB lost due to fragmentation)

  total time  =        8.37 secs   (8370 ticks @ 1000 us, 1 processor)
  total alloc = 218,627,008 bytes  (excludes profiling overheads)

to:

     418,734,144 bytes allocated in the heap
     169,549,048 bytes copied during GC
      40,965,384 bytes maximum residency (8 sample(s))
         385,384 bytes maximum slop
              69 MB total memory in use (0 MB lost due to fragmentation)

  total time  =        7.59 secs   (7586 ticks @ 1000 us, 1 processor)
  total alloc = 257,666,440 bytes  (excludes profiling overheads)

This shows a modest speedup and a large decrease in maximum memory
usage.  This indicates that earlier messages are being GC'd while
the list is still being iterated.  More heap allocation is taking
place, but we are recycling GC'd memory rather than claiming more
memory from the OS.