always pass ForeignPtrs to child objects
To avoid GC of parent causing errors when child objects (which are
opaque pointers) contain hidden references to parent, always pass
foreign pointers of parent objects to child objects.  Therefore the
parent cannot be garbage collected before the child.

This change might make it possible to avoid many instances of
detachPtr but this will need to be carefully checked and measured.
So it is left for future work.