do not run next action after non-success status
The 'status' helper function is supposed to scrutinise a 'Status'
and run the given "next" action upon success, or return 'f (Left e)'
on error.  The implementation was wrong - it was running the "next"
action for both cases.  This could e.g. cause segfault when 'status'
followed a double pointer-style constructor, and the next action was
to peek in the returned pointer.

Fix the issue by tightening the constraint to Applicative and using
'pure' in the error case.