use "safe" for notmuch_database_open
In notmuch_database_open() Xapian's "FlintLock" facility acquires an
exclusive (write) lock on the notmuch database.  To acquire the
lock, it calls fork(2) and the child uses fcntl(lockfd, F_SETLK, fl)
to acquire the lock (communicating the result to the parent process
via a pipe).  The lock is released by notmuch_database_close().

In a program that opens and closes database "sessions", if a new
database session is started (via notmuch_database_open()) before the
previous one has been closed, the child process will block while it
waits to acquire the lock.  And it will be waiting forever, because:

1. notmuch_database_open() is an "unsafe" FFI call

2. garbage collection never occurs during an "unsafe" FFI call

3. finalizers are invoked by the garbage collector

4. notmuch_database_close() is called from the DatabaseHandle
   finalizer

Change the notmuch_database_open() foreign import to be "safe".
This allows GC to run during a call to notmuch_database_open(),
releasing the lock, allowing the child process to acquire the lock
and return.

Fixes: https://github.com/purebred-mua/purebred/issues/468